+++ include/mirage/persistent_kernel/tasks/ampere/multitoken_paged_attention_split_kv.cuh
@@ -39,7 +39,7 @@ __device__ __forceinline__ void multitoken_paged_attention_split_kv_task_impl(
     int const *paged_kv_indptr_buffer_ptr,
     int const *paged_kv_indices_buffer_ptr,
     int const *paged_kv_last_page_len_buffer_ptr,
-    int request_id,
+    int16_t request_id,
     bool qk_norm,
     bool rope,
     void const *q_norm_weight_ptr,
